+++ scripts/readme/extract_steps_from_readme.py
@@ -6,11 +6,28 @@ from ghactions_driver.readme_parse import readme_parser
 from ghactions_driver.readme_step import ReadmeStepsManage
 
 
+def comment_pytest(full_text: str):
+    # separate full_text into lines, group them by start with pytest or not
+    full_text_lines = full_text.split("\n")
+    full_text_pytest = []
+    full_text_no_pytest = []
+    for line in full_text_lines:
+        if line.strip().startswith("python -m unittest"):
+            full_text_pytest.append(line)
+        else:
+            full_text_no_pytest.append(line)
+    return "\n".join(full_text_no_pytest), "\n".join(full_text_pytest)
+
+
 def write_readme_shell(readme_path: str, output_folder: str):
     full_text = readme_parser(readme_path)
+    full_text, full_text_pytest = comment_pytest(full_text)
     Path(ReadmeStepsManage.git_base_dir())
     bash_script_path = (
         Path(ReadmeStepsManage.git_base_dir()) / output_folder / "bash_script.sh"
+    )
+    bash_script_backup = (
+        Path(ReadmeStepsManage.git_base_dir()) / output_folder / "bash_script_pytest.sh"
     )
     template_env = Environment(
         loader=FileSystemLoader(
@@ -21,6 +38,9 @@ def write_readme_shell(readme_path: str, output_folder: str):
     bash_script_template = template_env.get_template("bash_script.sh.jinja2")
     with open(bash_script_path, "w") as f:
         f.write(bash_script_template.render({"command": full_text}))
+    if full_text_pytest != "":
+        with open(bash_script_backup, "w") as f:
+            f.write(bash_script_template.render({"command": full_text_pytest}))
 
 
 if __name__ == "__main__":