+++ src/promptflow-core/promptflow/core/_serving/response_creator.py
@@ -5,7 +5,7 @@
 import json
 import time
 from abc import ABC, abstractmethod
-from types import GeneratorType
+from types import AsyncGeneratorType, GeneratorType
 
 from promptflow.core._serving._errors import MultipleStreamOutputFieldsNotSupported, NotAcceptable
 
@@ -28,12 +28,15 @@ class ResponseCreator(ABC):
         response_original_value=False,
     ):
         # Fields that are with GeneratorType are streaming outputs.
-        stream_fields = [k for k, v in flow_run_result.items() if isinstance(v, GeneratorType)]
+        stream_fields = [
+            k for k, v in flow_run_result.items() if isinstance(v, GeneratorType) or isinstance(v, AsyncGeneratorType)
+        ]
         if len(stream_fields) > 1:
             raise MultipleStreamOutputFieldsNotSupported()
 
         self.stream_field_name = stream_fields[0] if stream_fields else None
         self.stream_iterator = flow_run_result.pop(self.stream_field_name, None)
+        self.is_async_streaming = isinstance(self.stream_iterator, AsyncGeneratorType)
         self.non_stream_fields = flow_run_result
 
         # According to RFC2616, if "Accept" header is not specified,
@@ -79,6 +82,24 @@ class ResponseCreator(ABC):
         # If there is stream field, read and yield data until the end.
         if self.stream_iterator is not None:
             for chunk in self.stream_iterator:
+                if self._on_stream_event:
+                    self._on_stream_event(chunk)
+                yield format_event({self.stream_field_name: chunk})
+        if self._on_stream_end:
+            duration = (time.time() - start_time) * 1000
+            self._on_stream_end(duration)
+
+    async def generate_async(self):
+        start_time = time.time()
+        if self._on_stream_start:
+            self._on_stream_start()
+        # If there are non streaming fields, yield them firstly.
+        if self.non_stream_fields:
+            yield format_event(self.non_stream_fields)
+
+        # If there is stream field, read and yield data until the end.
+        if self.stream_iterator is not None:
+            async for chunk in self.stream_iterator:
                 if self._on_stream_event:
                     self._on_stream_event(chunk)
                 yield format_event({self.stream_field_name: chunk})