+++ src/promptflow-devkit/promptflow/_sdk/operations/_flow_operations.py
@@ -43,7 +43,7 @@ from promptflow._sdk._utilities.general_utils import (
     copy_tree_respect_template_and_ignore_file,
     generate_flow_tools_json,
     generate_random_string,
-    generate_yaml_entry_without_recover,
+    generate_yaml_entry_without_delete,
     json_load,
     logger,
 )
@@ -108,7 +108,7 @@ class FlowOperations(TelemetryMixin):
         :rtype: dict
         """
         experiment = kwargs.pop("experiment", None)
-        flow = generate_yaml_entry_without_recover(entry=flow)
+        flow = generate_yaml_entry_without_delete(entry=flow)
         if Configuration.get_instance().is_internal_features_enabled() and experiment:
             if variant is not None or node is not None:
                 error = ValueError("--variant or --node is not supported experiment is specified.")
@@ -117,7 +117,7 @@ class FlowOperations(TelemetryMixin):
                     message=str(error),
                     error=error,
                 )
-            return self._client._experiments._test(
+            return self._client._experiments._test_flow(
                 flow=flow,
                 inputs=inputs,
                 environment_variables=environment_variables,
@@ -337,7 +337,7 @@ class FlowOperations(TelemetryMixin):
         """
         from promptflow._sdk._load_functions import load_flow
 
-        flow = generate_yaml_entry_without_recover(entry=flow)
+        flow = generate_yaml_entry_without_delete(entry=flow)
         flow = load_flow(flow)
         flow.context.variant = variant
 
@@ -1005,11 +1005,12 @@ class FlowOperations(TelemetryMixin):
             from promptflow.contracts.tool import ValueType
             from promptflow.core._model_configuration import PromptyModelConfiguration
 
-            flow_meta = {"inputs": entry._data.get("inputs", {})}
-            if "outputs" in entry._data:
-                flow_meta["outputs"] = entry._data.get("outputs")
-            elif include_primitive_output:
-                flow_meta["outputs"] = {"output": {"type": "string"}}
+            flow_meta = {
+                "inputs": entry._core_prompty._get_input_signature(),
+            }
+            output_signature = entry._core_prompty._get_output_signature(include_primitive_output)
+            if output_signature:
+                flow_meta["outputs"] = output_signature
             init_dict = {}
             for field in fields(PromptyModelConfiguration):
                 init_dict[field.name] = {"type": ValueType.from_type(field.type).value}