+++ src/promptflow/tests/test_configs/eager_flows/callable_class_with_openai/openai_chat.py
@@ -1,7 +1,6 @@
 from openai import AzureOpenAI
 
 from promptflow.connections import AzureOpenAIConnection
-from promptflow.tracing import trace
 
 
 class MyClass:
@@ -36,8 +35,7 @@ class MyClass:
                         yield chunk.choices[0].delta.content or ""
             # We must return the generator object, not using yield directly here.
             # Otherwise, the function itself will become a generator, despite whether stream is True or False.
-            # return generator()
-            return "".join(generator())
+            return generator()
         else:
             # chat api may return message with no content.
             return completion.choices[0].message.content or ""