+++ src/promptflow/tests/test_configs/flows/async_generator_tools/echo.py
@@ -1,0 +1,10 @@
+from promptflow.core import tool
+
+@tool
+async def echo(count):
+    """yield the input string."""
+
+    echo_text = "Echo - "
+    for i in range(count):
+        # yield a value to the caller
+        yield f"{echo_text}{i}"